+++ frameworks/ajax/system/request.js
@@ -535,7 +535,11 @@ SC.Request = SC.Object.extend(SC.Copyable, SC.Freezable,
     // Normalize arguments
     if (SC.typeOf(statusOrEvent) !== SC.T_NUMBER && SC.typeOf(statusOrEvent) !== SC.T_STRING) {
       // Accept multiple additional arguments (Do so before shifting the arguments!)
-      args = SC.A(arguments).slice(2);
+
+      // Fast arguments access.
+      // Accessing `arguments.length` is just a Number and doesn't materialize the `arguments` object, which is costly.
+      this.args = new Array(arguments.length - 2); //  SC.A(arguments).slice(2)
+      for (var i = 0, len = this.args.length; i < len; i++) { this.args[i] = arguments[i + 2]; }
 
       // Shift the arguments
       action = target;
@@ -543,7 +547,11 @@ SC.Request = SC.Object.extend(SC.Copyable, SC.Freezable,
       statusOrEvent = 0;
     } else {
       // Accept multiple additional arguments.
-      args = SC.A(arguments).slice(3);
+
+      // Fast arguments access.
+      // Accessing `arguments.length` is just a Number and doesn't materialize the `arguments` object, which is costly.
+      this.args = new Array(arguments.length - 3); //  SC.A(arguments).slice(3)
+      for (var i = 0, len = this.args.length; i < len; i++) { this.args[i] = arguments[i + 3]; }
     }
 
     // Prepare listeners for this object and notification target.