+++ frameworks/core_foundation/mixins/delegate_support.js
@@ -76,7 +76,11 @@ SC.DelegateSupport = {
     @returns {Object} value returned by delegate
   */
   invokeDelegateMethod: function(delegate, methodName, args) {
-    args = SC.A(arguments); args = args.slice(2, args.length) ;
+    // Fast arguments access.
+    // Accessing `arguments.length` is just a Number and doesn't materialize the `arguments` object, which is costly.
+    args = new Array(arguments.length - 2); //  SC.A(arguments).slice(2)
+    for (var i = 0, len = args.length; i < len; i++) { args[i] = arguments[i + 2]; }
+
     if (!delegate || !delegate[methodName]) delegate = this ;
 
     var method = delegate[methodName];