+++ frameworks/foundation/system/module.js
@@ -51,10 +51,16 @@ SC.Module = SC.Object.create(/** @scope SC.Module */ {
     @returns {Boolean} YES if already loaded, NO otherwise
   */
   loadModule: function (moduleName, target, method) {
-    var module = SC.MODULE_INFO[moduleName], callbacks, targets,
-        args   = SC.A(arguments).slice(3),
+    var module = SC.MODULE_INFO[moduleName],
+        //@if(debug)
         log    = SC.LOG_MODULE_LOADING,
-        idx, len;
+        //@endif
+        args;
+
+    // Fast arguments access.
+    // Accessing `arguments.length` is just a Number and doesn't materialize the `arguments` object, which is costly.
+    args = new Array(arguments.length - 3); //  SC.A(arguments).slice(3)
+    for (var i = 0, len = args.length; i < len; i++) { args[i] = arguments[i + 3]; }
 
     // Treat the first parameter as the callback if the target is a function and there is
     // no method supplied.